requirements.txt
appdirs==1.4.4
asgiref==3.2.7
async-lru==1.0.2
atomicwrites==1.4.0
attrs>=19.3.0
certifi>=2020.12.5
chardet>=4.0.0
colorama==0.4.4
cycler==0.10.0
distlib==0.3.1
Django>=3.0.5
django-filter==2.4.0
djangorestframework>=3.12.4
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
flake8==3.8.4
gunicorn==20.0.4
idna==2.9
iniconfig==1.1.1
isort==5.8.0
kiwisolver==1.3.1
matplotlib==3.3.4
mccabe==0.6.1
more-itertools==8.2.0
numpy==1.20.1
packaging==20.3
Pillow>=8.1.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py>=1.8.1
pyaes==1.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.1.0
pyparsing==2.4.7
Pyrogram==1.2.8
PySocks==1.7.1
pytest>=5.4.1
pytest-django>=4.2.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2019.3
requests>=2.25.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.2
urllib3>=1.25.9
virtualenv==20.4.2
wcwidth==0.1.9
zipp==3.1.0
